Dekrementierung des count-Wert in der Pivot-Tabelle bei löschen eines Artikels

+++ app/Http/Controllers/ShoppinglistController.php
@@ -156,12 +156,26 @@ class ShoppinglistController extends Controller
         $product_id = $request->input('product_id');
         $shoppinglist_id = $request->input('shoppinglist_id');
         $shoppinglist = Shoppinglist::find($shoppinglist_id);
-        $shoppinglist->products()->detach([$product_id]);
-        return response()->json([
-            'message' => 'Produkt erfolgreich gelöscht'
-        ]);
 
+        $product = $shoppinglist->products()->where('product_shoppinglist.product_id', '=', $product_id)->first();
+        $count = $product->pivot->count;
 
+        if($count > 1) {
+            $count--;
+            $shoppinglist->products()->updateExistingPivot($product->id, ['count' => $count]);
+            return response()->json([
+                'message' => 'Blablub',
+                'product_id' => $product->id,
+                'shoppinglist_id' => $shoppinglist->id,
+                'product_name' => $product->name,
+                'count' => $count,
+            ]);
+        } else {
+            $shoppinglist->products()->detach([$product_id]);
+            return response()->json([
+                'message' => 'Produkt erfolgreich gelöscht'
+            ]);
+        }
     }
 
     public function ajaxAddRecipe(Shoppinglist $shoppinglist)
